+++ src/test/kotlin/examples/kotlin/spring/canonical/CanonicalSpringKotlinTest.kt
@@ -512,7 +512,8 @@ open class CanonicalSpringKotlinTest {
 
     @Test
     fun testInsertSelect() {
-        val insertStatement = insertSelect(person) {
+        val insertStatement = insertSelect {
+            into(person)
             columns(id, firstName, lastName, birthDate, employed, occupation, addressId)
             select(add(id, constant<Int>("100")), firstName, lastName, birthDate, employed, occupation, addressId) {
                 from(person)
@@ -525,6 +526,56 @@ open class CanonicalSpringKotlinTest {
                 "select (id + 100), first_name, last_name, birth_date, employed, occupation, address_id " +
                 "from Person " +
                 "order by id"
+        )
+        val rows = template.insertSelect(insertStatement)
+        assertThat(rows).isEqualTo(6)
+
+        val records = template.select(id, firstName, lastName, birthDate, employed, occupation, addressId) {
+            from(person)
+            where { id isGreaterThanOrEqualTo 100 }
+            orderBy(id)
+        }.withRowMapper(personRowMapper)
+
+        assertThat(records).hasSize(6)
+        with(records[1]) {
+            assertThat(id).isEqualTo(102)
+            assertThat(firstName).isEqualTo("Wilma")
+            assertThat(lastName).isEqualTo(LastName("Flintstone"))
+            assertThat(birthDate).isNotNull
+            assertThat(employed).isTrue
+            assertThat(occupation).isEqualTo("Accountant")
+            assertThat(addressId).isEqualTo(1)
+        }
+    }
+
+    @Test
+    fun testInsertSelectNoTable() {
+        assertThatExceptionOfType(KInvalidSQLException::class.java).isThrownBy {
+            insertSelect {
+                columns(id, firstName, lastName, birthDate, employed, occupation, addressId)
+                select(add(id, constant<Int>("100")), firstName, lastName, birthDate, employed, occupation, addressId) {
+                    from(person)
+                    orderBy(id)
+                }
+            }
+        }.withMessage(Messages.getString("ERROR.29"))
+    }
+
+    @Test
+    fun testDeprecatedInsertSelect() {
+        val insertStatement = insertSelect(person) {
+            columns(id, firstName, lastName, birthDate, employed, occupation, addressId)
+            select(add(id, constant<Int>("100")), firstName, lastName, birthDate, employed, occupation, addressId) {
+                from(person)
+                orderBy(id)
+            }
+        }
+
+        assertThat(insertStatement.insertStatement).isEqualTo(
+            "insert into Person (id, first_name, last_name, birth_date, employed, occupation, address_id) " +
+                    "select (id + 100), first_name, last_name, birth_date, employed, occupation, address_id " +
+                    "from Person " +
+                    "order by id"
         )
         val rows = template.insertSelect(insertStatement)
         assertThat(rows).isEqualTo(6)